Use vh units instead of JavaScript to make table dimensions 50% of window height.
It should be [safe][1] to use viewport units, which means that the browser
can resize the table instead of us having to do it in JavaScript.
The viewport size isn't always exactly the same as the window size, but it
should have the same effect of making sure that the table is a
reasonable size.

[1]: http://caniuse.com/#search=vh